# Conflicts:
#	build.sbt
#	src/main/scala/euler/prob001_100/Prob072.scala
#	src/main/scala/euler/traits/UtilResult.scala
#	src/test/scala/euler/prob051_100Test.scala
#	src/test/scala/euler/prob701_800Test.scala

+++ src/main/scala/euler/traits/Util.scala
@@ -369,7 +369,7 @@ trait Util {
     if (n < 0)
       false
     else {
-      val tst: Long = (Math.sqrt(n) + 0.5).toLong
+      val tst: Long = (Math.sqrt(n.toDouble) + 0.5).toLong
       tst * tst == n
     }
   }
